refactor spot-related imports and remove unused zone repository
