--- vulcanus/src/main/java/com/merkury/vulcanus/model/repositories/ZoneRepository.java
@@ -1,9 +1,0 @@
-package com.merkury.vulcanus.model.repositories;
-
-import com.merkury.vulcanus.model.entities.Zone;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ZoneRepository extends JpaRepository<Zone, Long> {
-}
